% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaMMI-package.R
\docType{package}
\name{BaMMI-package}
\alias{BaMMI}
\alias{BaMMI-package}
\title{BaMMI: Bayesian Microbiome-Metabolite Integration}
\description{
A Bayesian framework for microbiome and metabolite integration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Kai-Jiang-1/BaMMI}
  \item Report bugs at \url{https://github.com/Kai-Jiang-1/BaMMI/issues}
}

}
\author{
\strong{Maintainer}: Kai Jiang \email{Kai.Jiang@uth.tmc.edu}

}
