% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malmo_plot_correlation.R
\name{malmo_plot_correlation}
\alias{malmo_plot_correlation}
\title{Correlation Heatmaps for MALMO}
\usage{
malmo_plot_correlation(Y, X)
}
\arguments{
\item{Y}{Numeric matrix of dimension n x q. Outcome matrix.
Can use \code{mydata$Y_true} or \code{mydata$Y_miss}.}

\item{X}{Numeric matrix of dimension n x p. Predictor matrix.
Can use \code{mydata$X}.}
}
\value{
A list containing:
\describe{
\item{cor_Y}{Numeric matrix of dimension q x q. Outcome vs
outcome correlation matrix.}
\item{cor_YX}{Numeric matrix of dimension q x p. Outcome vs
predictor correlation matrix.}
}
}
\description{
Produces two heatmaps: (1) outcome vs outcome correlation from
the observed data, and (2) outcome vs predictor correlation
from the observed data.
}
\examples{
\dontrun{
set.seed(123)
mydata <- gen_missing_value_mv(
  n = 50, p = 100, q = 3, snr = 1,
  proportion = 0.20, type = "mnar0.33"
)
malmo_plot_correlation(
  Y = mydata$Y_true,
  X = mydata$X
)
}
}
\seealso{
\code{\link{malmo_gibbs}}, \code{\link{malmo_summarize}}
}
